requirements.txt
behave
Faker
parse
parse-type
PyHamcrest
PyYAML
selenium